Migrate courses to backend API
