+++ back-end/test/server.test.js
@@ -19,4 +19,33 @@ describe("Server", () => {
         });
     });
   });
+
+  describe("GET /api/courses/:id", () => {
+    it("should return course details when course exists", (done) => {
+      chai
+        .request(app)
+        .get("/api/courses/1")
+        .end((err, res) => {
+          expect(res).to.have.status(200);
+          expect(res.body).to.include({
+            id: 1,
+            code: "CS 101",
+            title: "Intro to Computer Science",
+          });
+          expect(res.body).to.have.property("sections").that.is.an("array");
+          done();
+        });
+    });
+
+    it("should return 404 when course does not exist", (done) => {
+      chai
+        .request(app)
+        .get("/api/courses/999")
+        .end((err, res) => {
+          expect(res).to.have.status(404);
+          expect(res.body).to.have.property("error", "Course not found");
+          done();
+        });
+    });
+  });
 });
